auth_service/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
PyJWT==2.8.0
pydantic==2.3.0
flask-cors==5.0.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0